+++ packages/database/lib/src/database.dart
@@ -18,6 +18,7 @@ part 'daos/categories_dao.dart';
 part 'daos/budgets_dao.dart';
 part 'daos/import_sources_dao.dart';
 part 'daos/tags_dao.dart';
+part 'daos/recurring_dao.dart';
 
 /// Local finance database with all tables.
 @DriftDatabase(
@@ -47,6 +48,7 @@ class LocalFinanceDatabase extends _$LocalFinanceDatabase {
   late final BudgetsDao budgetsDao = BudgetsDao(this);
   late final ImportSourcesDao importSourcesDao = ImportSourcesDao(this);
   late final TagsDao tagsDao = TagsDao(this);
+  late final RecurringTransactionsDao recurringTransactionsDao = RecurringTransactionsDao(this);
 
   @override
   int get schemaVersion => 5;